Added search for category function in MovieDBDAO

+++ PrivateMovieCollection/src/privatemoviecollection/be/Rating.java
@@ -13,7 +13,7 @@ public class Rating
 {
 
     private int userRating;
-    private float IMDBRating;
+    private double IMDBRating;
 
     public int getUserRating()
     {
@@ -25,12 +25,12 @@ public class Rating
         this.userRating = userRating;
     }
 
-    public float getIMDBRating()
+    public double getIMDBRating()
     {
         return IMDBRating;
     }
 
-    public void setIMDBRating(float IMDBRating)
+    public void setIMDBRating(double IMDBRating)
     {
         this.IMDBRating = IMDBRating;
     }
